[Refactor] Introduce StorageBase module
