+++ be/module_boundary_manifest.json
@@ -402,6 +402,31 @@
       "remediation": "Keep StoragePrimitive limited to reusable storage contracts and value types; move concrete storage engine, tablet, rowset, lake, service, connector, and full Exec integration upward."
     },
     {
+      "id": "storagebase",
+      "doc_label": "StorageBase",
+      "summary": "Base storage algorithms and mask-buffer helpers above StoragePrimitive and ComputeEnv without concrete Storage engine, tablet, rowset, lake, service, or full Exec coupling.",
+      "owned_targets": ["StorageBase"],
+      "owned_roots": ["be/src/storage/base"],
+      "allowed_include_prefixes": [
+        "storage/base/",
+        "storage/primitive/",
+        "compute_env/sorting/",
+        "serde/",
+        "column/",
+        "types/",
+        "common/",
+        "base/",
+        "gutil/",
+        "gen_cpp/"
+      ],
+      "forbidden_include_prefixes": ["storage/", "compute_env/", "exec/", "service/", "connector/", "cache/", "http/", "agent/"],
+      "forbidden_includes": ["runtime/exec_env.h"],
+      "allowed_target_deps": ["ComputeEnv", "StoragePrimitive", "Serde", "RuntimeCore", "ChunkCore", "ColumnCore", "Types", "Common", "Base", "Gutil", "StarRocksGen"],
+      "allowed_test_targets": ["storage_base_test"],
+      "allowed_test_link_deps": ["StorageBase", "ComputeEnv", "StoragePrimitive", "Serde", "FSCore", "IO", "RuntimeCore", "ChunkCore", "ColumnCore", "Types", "Common", "Base", "Gutil", "StarRocksGen"],
+      "remediation": "Keep StorageBase limited to reusable storage algorithms and helpers that may depend on ComputeEnv; move concrete storage engine, tablet, rowset, lake, service, connector, cache, and full Exec integration upward."
+    },
+    {
       "id": "computeenv",
       "doc_label": "ComputeEnv",
       "summary": "Shared compute-side BE environment boundary for process-scoped compute resources and pipeline controls below full Exec/Storage and above RuntimeEnv.",
