+++ be/src/storage/lake/tablet_reader.cpp
@@ -33,17 +33,17 @@
 #include "runtime/env/global_env.h"
 #include "runtime/runtime_state.h"
 #include "storage/aggregate_iterator.h"
+#include "storage/base/merge_iterator.h"
+#include "storage/base/row_source_mask.h"
 #include "storage/chunk_helper.h"
 #include "storage/column_predicate_rewriter.h"
 #include "storage/conjunctive_predicates.h"
 #include "storage/lake/rowset.h"
 #include "storage/lake/utils.h"
 #include "storage/lake/versioned_tablet.h"
-#include "storage/merge_iterator.h"
 #include "storage/predicate_parser.h"
 #include "storage/primitive/empty_iterator.h"
 #include "storage/primitive/union_iterator.h"
-#include "storage/row_source_mask.h"
 #include "storage/rowset/rowid_range_option.h"
 #include "storage/rowset/rowset_options.h"
 #include "storage/rowset/short_key_range_option.h"
