+++ be/src/storage/load_chunk_spiller.cpp
@@ -25,13 +25,13 @@
 #include "runtime/runtime_state.h"
 #include "runtime/runtime_state_helper.h"
 #include "storage/aggregate_iterator.h"
+#include "storage/base/merge_iterator.h"
 #include "storage/chunk_helper.h"
 #include "storage/lake/tablet_internal_parallel_merge_task.h"
 #include "storage/lake/tablet_writer.h"
 #include "storage/load_spill_block_manager.h"
 #include "storage/load_spill_pipeline_merge_context.h"
 #include "storage/load_spill_pipeline_merge_iterator.h"
-#include "storage/merge_iterator.h"
 #include "storage/primitive/union_iterator.h"
 #include "storage/storage_metrics.h"
