+++ be/src/storage/load_spill_pipeline_merge_iterator.cpp
@@ -18,11 +18,11 @@
 #include "exec/spill/spiller.h"
 #include "exec/spill/spiller_factory.h"
 #include "runtime/runtime_state.h"
+#include "storage/base/merge_iterator.h"
 #include "storage/lake/tablet_internal_parallel_merge_task.h"
 #include "storage/lake/tablet_writer.h"
 #include "storage/load_chunk_spiller.h"
 #include "storage/load_spill_block_manager.h"
-#include "storage/merge_iterator.h"
 #include "storage/primitive/union_iterator.h"
 
 namespace starrocks {
