+++ be/src/storage/rowset/rowset.cpp
@@ -50,10 +50,10 @@
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
 #include "segment_options.h"
+#include "storage/base/merge_iterator.h"
 #include "storage/chunk_helper.h"
 #include "storage/delete_predicates.h"
 #include "storage/index/index_descriptor.h"
-#include "storage/merge_iterator.h"
 #include "storage/primitive/chunk_iterator.h"
 #include "storage/primitive/empty_iterator.h"
 #include "storage/primitive/projection_iterator.h"
