+++ be/src/storage/rowset/rowset_writer.cpp
@@ -59,14 +59,14 @@
 #include "segment_options.h"
 #include "serde/column_array_serde.h"
 #include "storage/aggregate_iterator.h"
+#include "storage/base/merge_iterator.h"
+#include "storage/base/row_source_mask.h"
 #include "storage/chunk_helper.h"
 #include "storage/index/index_descriptor.h"
-#include "storage/merge_iterator.h"
 #include "storage/metadata_util.h"
 #include "storage/olap_define.h"
 #include "storage/primitive/empty_iterator.h"
 #include "storage/primitive/type_utils.h"
-#include "storage/row_source_mask.h"
 #include "storage/rows_mapper.h"
 #include "storage/rowset/rowset.h"
 #include "storage/rowset/rowset_factory.h"
