+++ be/src/storage/rowset_merger.cpp
@@ -26,8 +26,8 @@
 #include "common/config_compaction_fwd.h"
 #include "common/config_exec_fwd.h"
 #include "gutil/stl_util.h"
+#include "storage/base/merge_iterator.h"
 #include "storage/chunk_helper.h"
-#include "storage/merge_iterator.h"
 #include "storage/primary_key_encoder.h"
 #include "storage/primitive/empty_iterator.h"
 #include "storage/primitive/union_iterator.h"
