+++ be/src/storage/tablet_reader.cpp
@@ -30,12 +30,12 @@
 #include "gutil/stl_util.h"
 #include "primary_key_encoder.h"
 #include "storage/aggregate_iterator.h"
+#include "storage/base/merge_iterator.h"
 #include "storage/chunk_helper.h"
 #include "storage/column_predicate.h"
 #include "storage/column_predicate_rewriter.h"
 #include "storage/conjunctive_predicates.h"
 #include "storage/delete_predicates.h"
-#include "storage/merge_iterator.h"
 #include "storage/olap_common.h"
 #include "storage/predicate_parser.h"
 #include "storage/primitive/empty_iterator.h"
